Move `uploadphoto` into `photo`
- Update paths and routes
- Keep `upload-errors` separate
