+++ routes.js
@@ -13,11 +13,10 @@ app.use('/prototype/filter', require('./routes/prototype/filter'));
 app.use('/prototype/apply', require('./routes/prototype/apply'));
 app.use('/prototype/intro', require('./routes/prototype/intro'));
 app.use('/prototype/photo', require('./routes/prototype/photo'));
+app.use('/prototype/photo/upload-errors', require('./routes/prototype/photo/upload-errors'));
 app.use('/prototype/help', require('./routes/prototype/help'));
 app.use('/prototype/photo-guide', require('./routes/prototype/photo-guide'));
 app.use('/prototype/photo-booth-shop', require('./routes/prototype/photo-booth-shop'));
-app.use('/prototype/photo/uploadphoto', require('./routes/prototype/photo/uploadphoto'));
-app.use('/prototype/photo/uploadphoto/upload-errors', require('./routes/prototype/photo/uploadphoto/upload-errors'));
 app.use('/prototype/takephoto', require('./routes/prototype/takephoto'));
 app.use('/prototype/filter-common-temp', require('./routes/prototype/filter-common-temp'));
 app.use('/prototype/startpage-temp', require('./routes/prototype/startpage-temp'));
