+++ routes/prototype/photo/upload-errors/steps.js
