--- routes/prototype/photo/uploadphoto/index.js
@@ -1,10 +1,0 @@
-var app = require('express')(),
-    wizard = require('hmpo-form-wizard'),
-    steps = require('./steps'),
-    fields = require('./fields');
-
-app.use(require('hmpo-template-mixins')(fields, { sharedTranslationKey: 'prototype' }));
-
-app.use(wizard(steps, fields, { templatePath: 'prototype/photo/uploadphoto', name: 'common' }));
-
-module.exports = app;
