+++ views/prototype/photo/final-checks.html
