+++ views/prototype/photo/processing-image.html
@@ -10,10 +10,10 @@
                 <script>
                 if (localStorage.journey === 'FTC') {
                     // Error journey
-                    // setTimeout(function(){window.location.href='../photo/uploadphoto/upload-errors/not-accepted'},1000);
-                    setTimeout(function(){window.location.href='../photo/uploadphoto/questions-intro'},1000);
+                    // setTimeout(function(){window.location.href='../photo/upload-errors/not-accepted'},1000);
+                    setTimeout(function(){window.location.href='../photo/questions-intro'},1000);
                 } else {
-                    setTimeout(function(){window.location.href='../photo/uploadphoto/questions-intro'},1000);
+                    setTimeout(function(){window.location.href='../photo/questions-intro'},1000);
                 }
                 </script>
             </div>
