+++ views/prototype/photo/questions-intro.html
@@ -13,11 +13,11 @@
 </div>
 <p>We may also check your photo manually. If something doesn't look right, we'll ask you for a different photo.</p>
 <br>
-<a href="/prototype/photo/uploadphoto/final-checks" class="button">Continue</a><br/><br/>
+<a href="/prototype/photo/final-checks" class="button">Continue</a><br/><br/>
 
 <div>
   <p>
-    <a href="../upload">Upload a different photo</a>
+    <a href="./upload">Upload a different photo</a>
   </p>
 </div>
 {{/content}}
