+++ views/prototype/photo/take-your-photo.html
