+++ views/prototype/photo/upload-errors/final-checks.html
