+++ views/prototype/photo/upload-errors/index.html
