+++ views/prototype/photo/upload-errors/not-accepted.html
