--- views/prototype/photo/uploadphoto/index.html
@@ -1,5 +1,0 @@
-{{< layout}}
-    {{$main-content}}
-
-    {{/main-content}}
-{{/ layout}}
